+++ packages/fresco-ui/src/form/components/fields/PasswordField.tsx
@@ -6,7 +6,10 @@ import { cx } from "../../../utils/cva";
 import { getPasswordStrength } from "./getPasswordStrength";
 import InputField from "./InputField";
 
-type PasswordFieldProps = React.ComponentProps<typeof InputField> & {
+type PasswordFieldProps = Omit<
+	React.ComponentProps<typeof InputField>,
+	"type"
+> & {
 	showStrengthMeter?: boolean;
 };